refactor sendPhoneCode mutation to setPhoneNumber
