+++ api-js/src/user/mutations/index.js
@@ -2,7 +2,7 @@ export * from './authenticate'
 export * from './reset-password'
 export * from './send-email-verification'
 export * from './send-password-reset'
-export * from './send-phone-code'
+export * from './set-phone-number'
 export * from './sign-in'
 export * from './sign-up'
 export * from './update-user-password'
